# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/users/register": {
            "post": {
                "description": "Register user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register an user with given information",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UserRegistrationForm"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/forms.UserRegistrationForm"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "forms.UserRegistrationForm": {
            "type": "object",
            "properties": {
                "is_student": {
                    "type": "boolean"
                },
                "nickname": {
                    "type": "string",
                    "example": "mengmota"
                },
                "partner_phone_number": {
                    "type": "string",
                    "example": "01012341234"
                },
                "password": {
                    "description": "password field",
                    "type": "string",
                    "example": "the!@#$pas123"
                },
                "phone_number": {
                    "description": "username field",
                    "type": "string",
                    "example": "01043214321"
                }
            }
        }
    }
}